Options!
Added options!

+++ ServiceWIM.ps1
@@ -1,11 +1,19 @@
 #Requires -Version 5.0
 #Requires -RunAsAdministrator
 #Requires -Modules Dism
-# Version 2.2
+# Version 3.0b
 
 # Variables
 	$MountDir = "$PSScriptRoot\Mount"
 	$Dismexe = "${env:ProgramFiles(x86)}\Windows Kits\10\Assessment and Deployment Kit\Deployment Tools\amd64\DISM\dism.exe"
+	$RunDefault = Read-Host 'Would you like to run with defaults?'
+
+# Check for Windows ADK
+	$ADKcheck = Test-Path $Dismexe
+	if ($ADKcheck -eq $False) {
+	Write-Output -InputObject "`nCannot find ADK DISM, exiting"
+	exit	
+	}
 	
 # Get vanilla WIM name
 	$origWIM = Read-Host 'What is the name of the current WIM file, including extension?'
@@ -22,41 +30,60 @@
 	& $Dismexe /Export-Image /SourceImageFile:$origWIM /SourceIndex:$sourceIndex /DestinationImageFile:$wim /Compress:Max /CheckIntegrity /quiet
 
 # Delete the old WIM
-	Write-Output -InputObject "`nDeleting original WIM file"
-	Set-ItemProperty -Path "$PSScriptRoot\$origWIM" -Name IsReadOnly -Value $false | Out-Null
-	Remove-Item $origWIM
+	if ($RunDefault -notlike "Y*") {
+		$deleteoldwim = Read-Host 'Would you like to remove the old WIM file?'
+	}
+	if ($deleteoldwim -like "Y*" -or $RunDefault -like "Y*") {
+		Write-Output -InputObject "`nDeleting original WIM file"
+		Set-ItemProperty -Path "$PSScriptRoot\$origWIM" -Name IsReadOnly -Value $false | Out-Null
+		Remove-Item $origWIM
 
 # Mount WIM file
 	Write-Output -InputObject "`nMounting WIM image"
 	if ((Test-Path $MountDir) -eq $False) {
-	New-Item -Path $PSScriptRoot -name "Mount" -ItemType "directory"
+		New-Item -Path $PSScriptRoot -name "Mount" -ItemType "directory"
 	}
 	& $Dismexe /Mount-Image /ImageFile:"$PSScriptRoot\$wim" /index:1 /MountDir:"$MountDir" /quiet
 
 # Apply default app associations
-	Write-Output -InputObject "`nSetting default App associations"
-	& $Dismexe /Image:"$MountDir" /Import-DefaultAppAssociations:"$PSScriptRoot\AppAssociations.xml" /quiet
+	if ($RunDefault -notlike "Y*") {
+		$ApplyDefaultApps = Read-Host 'Would you like to apply default app associations?'
+	}
+	if ($ApplyDefaultApps -like "Y*" -or $RunDefault -like "Y*") {
+		Write-Output -InputObject "`nSetting default app associations"
+		& $Dismexe /Image:"$MountDir" /Import-DefaultAppAssociations:"$PSScriptRoot\AppAssociations.xml" /quiet
+	}
 	
 # Disable features	
-	Write-Output -InputObject "`nDisabling Internet Explorer"
-	& $Dismexe /Image:"$MountDir" /Disable-Feature /FeatureName:Internet-Explorer-Optional-amd64 /quiet
+	if ($RunDefault -notlike "Y*") {
+		$RemoveIE = Read-Host 'Would you like to remove Internet Explorer?'
+	}
+	if ($RemoveIE -like "Y*" -or $RunDefault -like "Y*") {
+		Write-Output -InputObject "`nDisabling Internet Explorer"
+		& $Dismexe /Image:"$MountDir" /Disable-Feature /FeatureName:Internet-Explorer-Optional-amd64 /quiet
+	}
 
 # Install features
-	Write-Output -InputObject "`nInstalling .Net Framework 3.5 Feature"
-	if ($sourceVersion -like "10.0.18362*") {
-		& $Dismexe /image:"$MountDir" /enable-feature /featurename:NetFx3 /All /LimitAccess /Source:"$PSScriptRoot\net35" /quiet	
+	if ($RunDefault -notlike "Y*") {
+		$InstallNetF = Read-Host 'Would you like to install .Net Framework 3.5?'
 	}
+	if ($InstallNetF -like "Y*" -or $RunDefault -like "Y*") {
+		Write-Output -InputObject "`nInstalling .Net Framework 3.5 Feature"
+		if ($sourceVersion -like "10.0.18362*") {
+			& $Dismexe /image:"$MountDir" /enable-feature /featurename:NetFx3 /All /LimitAccess /Source:"$PSScriptRoot\net35" /quiet	
+		}
 
-	elseif ($sourceVersion -like "10.0.17763*") {
-		& $Dismexe /image:"$MountDir" /enable-feature /featurename:NetFx3 /All /LimitAccess /Source:"$PSScriptRoot\net35_1809" /quiet	
-	}
-	
-	else {
-		Write-Output -InputObject "`nVersion not detected, skipping .NET Framework install"
+		elseif ($sourceVersion -like "10.0.17763*") {
+			& $Dismexe /image:"$MountDir" /enable-feature /featurename:NetFx3 /All /LimitAccess /Source:"$PSScriptRoot\net35_1809" /quiet	
+		}
 	}
 
 # Customise registry
-	Write-Output -InputObject "`nCustomising registry"
+	if ($RunDefault -notlike "Y*") {
+		$CustomiseREG = Read-Host 'Would you like to customise the registry?'
+	}
+	if ($CustomiseREG -like "Y*" -or $RunDefault -like "Y*") {
+		Write-Output -InputObject "`nCustomising registry"
 
 	# Mount registry
 		REG LOAD "HKLM\_NTUSER" "$MountDir\Users\Default\NTUSER.DAT" | Out-Null
@@ -103,11 +130,20 @@
 # Unmount registry
 	REG UNLOAD "HKLM\_NTUSER" | Out-Null
 	REG UNLOAD "HKLM\_SOFTWARE" | Out-Null
+	}
 
 # Copy folders to local disk
-	Write-Output -InputObject "`nCopying Folders"
-	xcopy Root_Folders\*.* "$MountDir\" /EXCLUDE:CopyExclusions.txt /E /C /H
-
+	if ($RunDefault -notlike "Y*") {
+		$CopyRoot = Read-Host 'Would you like to copy folder from "Root_Folders" to the root of your WIM??'
+	}
+	if ($CopyRoot -like "Y*" -or $RunDefault -like "Y*") {
+		Write-Output -InputObject "`nCopying Folders"
+		xcopy Root_Folders\*.* "$MountDir\" /EXCLUDE:CopyExclusions.txt /E /C /H
+	}
 # Commit image
 	Write-Output -InputObject "`nCommiting WIM image, this may take some time"
 	& $Dismexe /Unmount-Image /MountDir:"$MountDir" /Commit
+
+# The end
+	Write-Output -InputObject "`nThe WIM image should now be commited with all chosen modifications, we've paused here so you can see any errors!"
+	pause
